Some adjustements and suggestions from melo
