+++ src/commands/mod.rs
@@ -1,0 +1,1 @@
+// this will eventually gets filled once commands are gonna be implemented.
