+++ src/events/mod.rs
@@ -1,0 +1,3 @@
+mod on_guild_member_add;
+
+pub use self::on_guild_member_add::handle as on_member_join;
